docs/examples/imageclass-wasm/requirements.txt
torch_to_nnef
torchvision==0.22.0
